+++ src/react/src/components/ProductionLocation/Sidebar/SupplyChain/SupplyChainNetworkDrawer/AnonymizedSection/constants.js
@@ -1,0 +1,3 @@
+const ANONYMIZED_SECTION_TITLE = 'Anonymized Data Sources';
+
+export default ANONYMIZED_SECTION_TITLE;