+++ src/react/src/components/ProductionLocation/Sidebar/SupplyChain/SupplyChainNetworkDrawer/DrawerHeader/constants.js
@@ -1,0 +1,3 @@
+const DRAWER_TITLE = 'All Data Sources';
+
+export default DRAWER_TITLE;