+++ src/react/src/components/ProductionLocation/Sidebar/SupplyChain/SupplyChainNetworkDrawer/DrawerSubtitle/styles.js
@@ -1,0 +1,13 @@
+import COLOURS from '../../../../../../util/COLOURS';
+
+export default () =>
+    Object.freeze({
+        subtitle: Object.freeze({
+            marginTop: '8px',
+            marginBottom: '0px',
+            fontSize: '1rem',
+            lineHeight: 1.4,
+            color: COLOURS.DARK_GREY,
+            paddingBottom: '16px',
+        }),
+    });