+++ src/react/src/components/ProductionLocation/Sidebar/SupplyChain/SupplyChainNetworkDrawer/PublicContributorsSection/constants.js
@@ -1,0 +1,3 @@
+const UPLOADED_VIA_LIST_LABEL = 'Uploaded via list';
+
+export default UPLOADED_VIA_LIST_LABEL;