+++ src/react/src/components/ProductionLocation/Sidebar/SupplyChain/constants.js
@@ -1,0 +1,15 @@
+const PLURAL_MAP = {
+    'Academic / Researcher / Journalist / Student':
+        'Academics / Researchers / Journalists / Students',
+    'Auditor / Certification Scheme / Service Provider':
+        'Auditors / Certification Schemes / Service Providers',
+    'Brand / Retailer': 'Brands / Retailers',
+    'Civil Society Organization': 'Civil Society Organizations',
+    'Facility / Factory / Manufacturing Group / Supplier / Vendor':
+        'Facilities / Factories / Manufacturing Groups / Suppliers / Vendors',
+    'Multi-Stakeholder Initiative': 'Multi-Stakeholder Initiatives',
+    Union: 'Unions',
+    Other: 'Others',
+};
+
+export default PLURAL_MAP;